requirements.txt
streamlit
ccxt
pandas
numpy
plotly
python-dotenv